+++ apps/block_scout_web/lib/block_scout_web/views/api/v2/block_view.ex
@@ -3,7 +3,6 @@ defmodule BlockScoutWeb.API.V2.BlockView do
 
   alias BlockScoutWeb.BlockView
   alias BlockScoutWeb.API.V2.{ApiView, Helper}
-  alias Explorer.Chain
   alias Explorer.Chain.Block
   alias Explorer.Counters.BlockPriorityFeeCounter
 
@@ -29,10 +28,10 @@ defmodule BlockScoutWeb.API.V2.BlockView do
   end
 
   def prepare_block(block, _conn, single_block? \\ false) do
-    burned_fee = Chain.burned_fees(block.transactions, block.base_fee_per_gas)
+    burnt_fees = Block.burnt_fees(block.transactions, block.base_fee_per_gas)
     priority_fee = block.base_fee_per_gas && BlockPriorityFeeCounter.fetch(block.hash)
 
-    tx_fees = Chain.txn_fees(block.transactions)
+    transaction_fees = Block.transaction_fees(block.transactions)
 
     %{
       "height" => block.number,
@@ -48,7 +47,7 @@ defmodule BlockScoutWeb.API.V2.BlockView do
       "gas_limit" => block.gas_limit,
       "nonce" => block.nonce,
       "base_fee_per_gas" => block.base_fee_per_gas,
-      "burnt_fees" => burned_fee,
+      "burnt_fees" => burnt_fees,
       "priority_fee" => priority_fee,
       # "extra_data" => "TODO",
       "uncles_hashes" => prepare_uncles(block.uncle_relations),
@@ -56,9 +55,9 @@ defmodule BlockScoutWeb.API.V2.BlockView do
       "rewards" => prepare_rewards(block.rewards, block, single_block?),
       "gas_target_percentage" => gas_target(block),
       "gas_used_percentage" => gas_used_percentage(block),
-      "burnt_fees_percentage" => burnt_fees_percentage(burned_fee, tx_fees),
+      "burnt_fees_percentage" => burnt_fees_percentage(burnt_fees, transaction_fees),
       "type" => block |> BlockView.block_type() |> String.downcase(),
-      "tx_fees" => tx_fees,
+      "tx_fees" => transaction_fees,
       "withdrawals_count" => count_withdrawals(block)
     }
     |> chain_type_fields(block, single_block?)
@@ -105,8 +104,9 @@ defmodule BlockScoutWeb.API.V2.BlockView do
 
   def burnt_fees_percentage(_, %Decimal{coef: 0}), do: nil
 
-  def burnt_fees_percentage(burnt_fees, tx_fees) when not is_nil(tx_fees) and not is_nil(burnt_fees) do
-    burnt_fees.value |> Decimal.div(tx_fees) |> Decimal.mult(100) |> Decimal.to_float()
+  def burnt_fees_percentage(burnt_fees, transaction_fees)
+      when not is_nil(transaction_fees) and not is_nil(burnt_fees) do
+    burnt_fees.value |> Decimal.div(transaction_fees) |> Decimal.mult(100) |> Decimal.to_float()
   end
 
   def burnt_fees_percentage(_, _), do: nil