+++ apps/block_scout_web/lib/block_scout_web/views/wei_helper.ex
@@ -53,7 +53,11 @@ defmodule BlockScoutWeb.WeiHelper do
       "10"
   """
   @spec format_wei_value(Wei.t(), Wei.unit(), format_options()) :: String.t()
-  def format_wei_value(%Wei{} = wei, unit, options \\ []) when unit in @valid_units do
+  def format_wei_value(_wei, _unit, _options \\ [])
+
+  def format_wei_value(nil, _unit, _options), do: nil
+
+  def format_wei_value(%Wei{} = wei, unit, options) when unit in @valid_units do
     converted_value =
       wei
       |> Wei.to(unit)