+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 
 - [#8955](https://github.com/blockscout/blockscout/pull/8955) - Remove daily balances updating from BlockReward fetcher
+- [#8846](https://github.com/blockscout/blockscout/pull/8846) - Handle nil gas_price at address view
 
 ### Chore